added OpenAPI support; added JWT verification on each critical endpoint (you can't delete or view a file if it doesn't belong to you); moved the JWT Secret Key to a .env file that is NOT pushed (consult README.md to see how to implement it back); minor frontend fixes; minor tests fixes

+++ api/src/main/java/com/openclassrooms/dataShare_api/controller/UserController.java
@@ -8,6 +8,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -25,6 +27,10 @@ import com.openclassrooms.dataShare_api.model.User;
 import com.openclassrooms.dataShare_api.service.FileService;
 import com.openclassrooms.dataShare_api.service.UserService;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.persistence.EntityExistsException;
 import lombok.extern.slf4j.Slf4j;
 
@@ -46,16 +52,44 @@ public class UserController {
      * @param id
      * @return [the User, 200 OK]
      */
+    @Operation(
+        summary = "Fetches a given User"
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "User found"),
+        @ApiResponse(responseCode = "404", description = "User not found"),
+        @ApiResponse(responseCode = "401", description = "Unauthorized"),
+        @ApiResponse(responseCode = "500", description = "Incorrect JWT provided")
+    })
+    @SecurityRequirement(name = "bearerAuth")
     @GetMapping("/user/{id}")
-    public ResponseEntity<User> getUser(@PathVariable Long id) {
-        log.info("[GET] /api/user/" + id + " (" + HttpStatus.OK + ")");
-        return ResponseEntity.ok(userService.getUser(id));
+    public ResponseEntity<User> getUser(@PathVariable Long id, @AuthenticationPrincipal Jwt jwt) {
+        Long ownerId = jwt.getClaim("userId");
+
+        try {
+            if (!id.equals(ownerId))
+                throw new Exception("Incorrect user token provided.");
+
+            log.info("[GET] /api/user/" + id + " (" + HttpStatus.OK + ")");
+            return ResponseEntity.ok(userService.getUser(id));
+        } catch (Exception e) {
+            log.error("[GET] /api/user/" + id + "(" + HttpStatus.INTERNAL_SERVER_ERROR + "): " + e.toString());
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     /**
      * Fetches every User
      * @return [a list of every User, 200 OK]
      */
+    @Operation(
+        summary = "Fetches every registered User"
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "Users found"),
+        @ApiResponse(responseCode = "401", description = "Unauthorized")
+    })
+    @SecurityRequirement(name = "bearerAuth")
     @GetMapping("/users")
     public ResponseEntity<List<User>> getUsers() {
         log.info("[GET] /api/users (" + HttpStatus.OK + ")");
@@ -68,9 +102,24 @@ public class UserController {
      * @param user the new data
      * @return 200 OK if successful; 404 NOT_FOUND if the User is not found; 500 INTERNAL_SERVER_ERROR otherwise
      */
+    @Operation(
+        summary = "Updates a given User"
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "User updated"),
+        @ApiResponse(responseCode = "404", description = "User not found"),
+        @ApiResponse(responseCode = "401", description = "Unauthorized"),
+        @ApiResponse(responseCode = "500", description = "Incorrect JWT provided")
+    })
+    @SecurityRequirement(name = "bearerAuth")
     @PutMapping("/user/{id}")
-    public ResponseEntity<HttpStatusCode> updateUser(@PathVariable Long id, @Validated @RequestBody User user) {
+    public ResponseEntity<HttpStatusCode> updateUser(@PathVariable Long id, @Validated @RequestBody User user, @AuthenticationPrincipal Jwt jwt) {
+        Long ownerId = jwt.getClaim("userId");
+        
         try {
+            if (!id.equals(ownerId))
+                throw new Exception("Incorrect user token provided");
+
             userService.updateUser(id, user);
             log.info("[PUT] /api/user (" + HttpStatus.OK + ")");
             return new ResponseEntity<>(HttpStatus.OK);
@@ -88,9 +137,24 @@ public class UserController {
      * @param id
      * @return 200 OK if successful; 404 NOT_FOUND if the User is not found; 500 INTERNAL_SERVER_ERROR if another error occurred (for instance, during the User's files deletion process)
      */
+    @Operation(
+        summary = "Deletes a given User"
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "User deleted"),
+        @ApiResponse(responseCode = "404", description = "User not found"),
+        @ApiResponse(responseCode = "401", description = "Unauthorized"),
+        @ApiResponse(responseCode = "500", description = "Incorrect JWT provided; Error during a User's Files deletion process")
+    })
+    @SecurityRequirement(name = "bearerAuth")
     @DeleteMapping("/user/{id}")
-    public ResponseEntity<HttpStatusCode> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<HttpStatusCode> deleteUser(@PathVariable Long id, @AuthenticationPrincipal Jwt jwt) {
+        Long ownerId = jwt.getClaim("userId");
+
         try {
+            if (!id.equals(ownerId))
+                throw new Exception("Incorrect user token provided");
+
             //To delete the User, we have to first delete his files
             for (DSFile file : fileService.getFiles(id))
                 fileService.deleteFile(file.getId());
@@ -117,6 +181,14 @@ public class UserController {
      * @param loginDTO the login information (email + password)
      * @return [a valid Jwt, 200 OK]; 404 NOT_FOUND if the User is not found; 401 UNAUTHORIZED if the password is incorrect; 500 INTERNAL_SERVER_ERROR otherwise
      */
+    @Operation(
+        summary = "Logs in a given User"
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "Credentials validated"),
+        @ApiResponse(responseCode = "404", description = "User not found"),
+        @ApiResponse(responseCode = "401", description = "Credentials invalid")
+    })
     @PostMapping("/login")
     public ResponseEntity<TokenDTO> login(@RequestBody LoginDTO loginDTO) {
         try {
@@ -141,6 +213,14 @@ public class UserController {
      * @param user the register information (email + password)
      * @return 201 CREATED if successful; 400 BAD_REQUEST if a user with that email already exists or the provided information is invalid; 500 INTERNAL_SERVER_ERROR otherwise
      */
+    @Operation(
+        summary = "Registers a new User"
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "201", description = "Account created"),
+        @ApiResponse(responseCode = "400", description = "Email/password invalid; Account with same email already exists"),
+        @ApiResponse(responseCode = "500", description = "Incorrect JWT provided")
+    })
     @PostMapping("/register")
     public ResponseEntity<HttpStatusCode> register(@Validated @RequestBody User user) {
         try {
